实现通过价格区间搜索TODO：分页

+++ src/main/java/com/chr/website/mapper/productMapper.java
@@ -53,6 +53,6 @@ public interface productMapper {
      */
     List<product> search(@Param("keyword") String keyword,
                          @Param("CategoryId") Integer CategoryId,
-                         @Param("min") Integer min,
-                         @Param("max") Integer max);
+                         @Param("min") Double min,
+                         @Param("max") Double max);
 }
